feat(flagship): lifecycle hooks
Right now flagship's module patching scripts can only apply to a specific module in a fixed event (ios before link and android after link). And we cannot add project specific patches easily using current framework.

Replaced the current design with a more extensible lifecycle hook design. So a user can:
- Add project specific hooks by creating `./buildHooks` folder under project root path
- Execute the script in any of supported lifecycle events. Currently supporting: 'beforeCopyBoilerplate' | 'beforeLink' | 'beforeIOSPodInstall' | 'afterLink' | 'failed';
- Filter hooks by platforms / dependencies / dependency versions

No breaking change - all the current patches are restructured in a way that behaves exactly same as the previous commit. Still we should scrutiny all the current patches to make them more restrictive though (add dependency version conditions, etc)

- Revisit exiting patches to apply them more appropriate conditions
- Split some global fixes into separate hooks to make them more modular
- Add more lifecycle events if needed
