+++ packages/flagship/package.json
@@ -20,6 +20,7 @@
     "react-native": "^0.55.4",
     "react-native-svg": "^7.0.2",
     "replace-in-file": "^3.4.0",
+    "semver": "^5.6.0",
     "yargs": "^12.0.1"
   },
   "rnpm": {
@@ -32,6 +33,7 @@
   },
   "devDependencies": {
     "@brandingbrand/fsengage": "2.0.0",
+    "@types/semver": "^5.5.0",
     "@types/yargs": "^12.0.1",
     "true-case-path": "^1.0.3"
   },
